// src/shared/lib/rental-offer-generator/tsv-rental-offer.enum.ts
export enum GenerateMockCountPassword {
  MinValue = 1,
  MaxValue = 10,
}

export enum GenerateMockCommentsCount {
  MinValue = 1,
  MaxValue = 10,
}

export enum GenerateMockRatingCount {
  MinValue = 1,
  MaxValue = 5,
}

export enum GenerateMockRoomsCount {
  MinValue = 1,
  MaxValue = 8,
}

export enum GenerateMockGuestsCount {
  MinValue = 1,
  MaxValue = 10,
}

export enum GenerateMockPriceCount {
  MinValue = 100,
  MaxValue = 100000,
}